Don't require GITREVUE_TOKEN to always be present. The user may not be interacting with the API. (e.g --dry-run, --version, or --help)
