+++ src/utils.ts
@@ -1,5 +1,6 @@
 import fs from 'fs'
 import glob from 'glob'
+import { Config } from "@gitrevue/sdk";
 
 export const resolveGlobs = (paths: string[]): string[] => {
     return paths.reduce((files: string[], path: string): string[] => {
@@ -13,4 +14,17 @@ export const getSize = (path: string): number => {
     const stats = fs.statSync(path)
 
     return stats.size
+}
+
+export const getConfig = (): Config => {
+    const config: Config = {
+        token: process.env.GITREVUE_TOKEN,
+        url: (process.env.GITREVUE_URL || 'https://gitrevue.io').replace(/\/+$/gm, ''),
+    }
+
+    if (!config.token) {
+        throw new Error('Unable to detect authorization token. Please specify your token using the GITREVUE_TOKEN environment variable. You can create a token at http://gitrevue.test/settings#/api')
+    }
+
+    return config
 }
